split methods to another class to pass CI test
